Restructure api directory
