+++ backend/api/middlewares/cors_middleware.py
@@ -1,0 +1,17 @@
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+
+
+def add_cors_middleware(app: FastAPI) -> None:
+    """
+    Adds the cors middleware to the FastAPI app
+
+    @param app: FastAPI app to add the middleware to
+    """
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["http://localhost:*"],
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
